+++ src/index.ts
@@ -44,7 +44,12 @@ export { default as Social, type SocialProps } from './patterns/Social/Social';
 export { default as ViewingsList, type ViewingsListProps } from './patterns/ViewingsList/ViewingsList';
 export { default as Modal, type ModalProps } from './components/Modal/Modal';
 export { default as Drawer } from './components/Drawer/Drawer';
-export { default as Pagination, type PaginationProps } from './components/Pagination/Pagination';
+export {
+  default as Pagination,
+  type PaginationProps,
+  type PaginationOption,
+  type PaginationOptionValue,
+} from './components/Pagination/Pagination';
 
 export {
   default as StatefulViewingsList,
@@ -52,7 +57,6 @@ export {
 } from './patterns/ViewingsList/StatefulViewingsList';
 export * from './components/Text';
 export * from './components/TextSymbol';
-// export* from './components/HTMLParser';
 export * from './components/Accordion';
 export { default as UserManagement, type UserManagementProps } from './patterns/UserManagement/UserManagement';
 export { AuthState } from './patterns/UserManagement/types';
@@ -66,3 +70,6 @@ export * from './components/Collapsible';
 export * from './providers/SeldonProvider';
 export { default as PageContentWrapper } from './components/PageContentWrapper/PageContentWrapper';
 export * from './components/Carousel';
+export * from './components/Detail';
+export * from './patterns/DetailList';
+export * from './components/PinchZoom';