+++ src/patterns/DetailList/types.ts
@@ -1,0 +1,16 @@
+export enum DetailListAlignment {
+  /*
+   * Aligns each Detail's label and value in columns.
+   * | Label1   Value1           |
+   * | Label2   Value2           |
+   * | Label3   Value3           |
+   */
+  columns = 'columns',
+  /*
+   * Justifies each Detail's label and value to the outside of the container
+   * | Label1             Value1 |
+   * | Label2             Value2 |
+   * | Label3             Value3 |
+   */
+  justified = 'justified',
+}